Changing the default initial table size to 256 almost works reliably on Windows10.  At least within a virtual machine there is a race condition that either occasionally|always produces messages like:
  FTab.newTab cannot replace old index "..": The process cannot access
  the file because it is being used by another process

out of test/basic.nim with various brand new file bases & ob counts.

The only process is itself and it has just closed the file.  Maybe some
system thread holds a lock for a short while and a loop would fix this?
There is also a chance I am not aware of some more fundamental problem.
We need `MoveFileTransacted` (and so NTFS) for safe parallel reader
processes anyway, but that may come with its own requirements.